Se agregó: anotaciones Entity y Table a las clases. Cambios en las capas Service-Controller

+++ Estrada_Rodolfo_3987_TP4/src/main/java/ar/edu/unju/fi/tp4/services/IProductoService.java
@@ -9,7 +9,7 @@ public interface IProductoService {
 	public void addProducto(Producto producto);
 	public Producto getUltimoProducto();
 	public Producto getProducto();
-	public Producto findByCodigo(int codigo);
+	public Producto findByCodigo(Long codigo);
 	public void generarListProductos();
 	public List<Producto> getListProductos();
